+++ geo/CHANGES.md
@@ -2,13 +2,19 @@
 
 ## Unreleased
 
+* Bump proj crate to 0.23.0, using PROJ 8.1.0
 * Add `ChaikinSmoothing` algorithm
 * Fix `rotate` for multipolygons to rotate around the collection's centroid, instead of rotating each individual polygon around its own centroid.
   * <https://github.com/georust/geo/pull/651>
 * Add KNearestConcaveHull algorithm
   * <https://github.com/georust/geo/pull/635>
+<<<<<<< HEAD
 * Implemented `ClosestPoint` method for types Triangle, Rect, GeometryCollection, Coordinate and the Geometry enum.
   * <https://github.com/georust/geo/pull/675>
+=======
+* remove cargo-tarpaulin due to instability (#676, #677)
+* Fix: `ClosestPoint` for Polygon's handling of internal points
+>>>>>>> b5d1a7f6b805ce12224390076fa338fb460e4632
 
 ## 0.18.0